requirements.txt
pandas==2.2.3
pendulum==3.1.0
minio==7.2.15
cryptography==45.0.2
apache-airflow-providers-slack==8.9.0
pangres==4.2.1